perf: connect WebSocket client to VS Code extension ~5s earlier
- Make websocketClient.ts a leaf module: import CONSTANTS from a new
  dependency-free config/constants.ts, inline isCI, defer uxLog imports.
  Importing the client now loads ~215 modules in ~0.1s instead of ~1500
  modules in ~3.4s (langchain, puppeteer, jira.js... via the utils barrel)
- start-ws-client hook: connect BEFORE loading the command class (class
  loading can take up to 2s); resolve plugin root from the manifest only;
  inline isAgentMode so the utils barrel is never loaded pre-connect
- Reuse SFDX_HARDIS_COMMAND_CONTEXT_ID (sent by the VS Code extension) as
  websocket context id so the extension can open the command panel at click
  time and this client adopts it on connection
- check-upgrade hook: return immediately when disabled (env checks before
  any import) and run the whole check fire-and-forget so it never delays
  the WebSocket connection or command startup
- Replace the 250ms polling in WebSocketClient.isInitialized() with an
  event-driven promise resolved on the extension's userInput answer
